refactor: Move shared components to parent folder
